+++ apis/core/bootstrap/shapes/index.ts
@@ -4,7 +4,7 @@ import { DatasetShape } from './dataset'
 import { CSVSourceShape } from './csv-source'
 import { JobUpdateShape, JobTriggerShape } from './jobs'
 import { ColumnMappingShape } from './column-mapping'
-import { DimensionMetadataShape, ManagedDimensionMappingShape } from './dimension'
+import { DimensionMetadataShape, SharedDimensionMappingShape } from './dimension'
 
 export default [
   CubeProjectShape,
@@ -15,5 +15,5 @@ export default [
   JobUpdateShape,
   JobTriggerShape,
   DimensionMetadataShape,
-  ManagedDimensionMappingShape,
+  SharedDimensionMappingShape,
 ]