+++ apis/core/lib/domain/dimension/update.ts
@@ -7,7 +7,7 @@ import { prov, rdf, schema } from '@tpluscode/rdf-ns-builders'
 import { ResourceStore } from '../../ResourceStore'
 import * as id from '../identifiers'
 import { findProject } from '../queries/cube-project'
-import { canBeMappedToManagedDimension } from './DimensionMetadata'
+import { canBeMappedToSharedDimension } from './DimensionMetadata'
 
 interface UpdateDimensionCommand {
   metadataCollection: NamedNode
@@ -41,7 +41,7 @@ export async function update({
 
   const dimension = metadata.updateDimension(dimensionMetadata)
 
-  if (canBeMappedToManagedDimension(dimension) && !dimension.mappings) {
+  if (canBeMappedToSharedDimension(dimension) && !dimension.mappings) {
     const project = await store.getResource<Project>(await findProject(metadata))
 
     const slug = dimension.id.value.substring(dimension.id.value.lastIndexOf('/') + 1)
@@ -51,7 +51,7 @@ export async function update({
       .addOut(schema.about, dimension.about)
 
     dimension.mappings = dimensionMapping.term
-  } else if (!canBeMappedToManagedDimension(dimension) && dimension.mappings) {
+  } else if (!canBeMappedToSharedDimension(dimension) && dimension.mappings) {
     store.delete(dimension.mappings)
     dimension.mappings = undefined
   }