+++ ui/CHANGELOG.md
@@ -7,7 +7,7 @@
 - f36c9a9: Improve form validation errors (#175)
 - 8726055: Add link to documentation (#545)
 - b1d57e4: Fix CSV upload when file name contains special characters (#513)
-- eaf8e00: User interface to map a dimension to a managed dimension
+- eaf8e00: User interface to map a dimension to a shared dimension
 - d4d2409: Disable OIDC session monitoring, which sometimes causes an incomplete user login information (#480)
 
 ## 0.1.16